+++ include/wx/msw/textctrl.h
@@ -61,7 +61,7 @@ public:
     void Clear() override;
 
     int GetLineLength(long lineNo) const override;
-    std::string GetLineText(long lineNo) const override;
+    std::string GetLineText(int lineNo) const override;
     int GetNumberOfLines() const override;
 
     void SetMaxLength(unsigned long len) override;
